[mod/#119]addsports->my로 패키지 구조 변경
