+++ app/src/main/java/com/smashing/app/data/di/my/MyRepositoryModule.kt
@@ -1,4 +1,4 @@
-package com.smashing.app.data.di.profile.myprofile
+package com.smashing.app.data.di.my
 
 import com.smashing.app.data.repository.api.MyRepository
 import com.smashing.app.data.repository.impl.MyRepositoryImpl
