+++ app/src/main/java/com/smashing/app/data/remote/datasource/api/MyRemoteDataSource.kt
@@ -2,12 +2,15 @@ package com.smashing.app.data.remote.datasource.api
 
 
 import com.smashing.app.data.remote.dto.BaseResponse
-import com.smashing.app.data.remote.dto.profile.my.MyPageData
-import com.smashing.app.data.remote.dto.profile.my.MyProfileReviewListData
+import com.smashing.app.data.remote.dto.my.AddSportProfileRequest
+import com.smashing.app.data.remote.dto.my.MyPageData
+import com.smashing.app.data.remote.dto.my.MyProfileReviewListData
 
 interface MyRemoteDataSource {
     suspend fun getMyProfile(): BaseResponse<MyPageData>
 
     suspend fun getMyGameReviews(cursor: String?, size: Int?): BaseResponse<MyProfileReviewListData>
+
+    suspend fun addSportProfile(request: AddSportProfileRequest): BaseResponse<String?>
 
 }
