--- app/src/main/java/com/smashing/app/data/remote/service/AddSportsService.kt
@@ -1,14 +1,0 @@
-package com.smashing.app.data.remote.service
-
-import com.smashing.app.data.remote.dto.BaseResponse
-import com.smashing.app.data.remote.dto.addsports.AddSportProfileRequest
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface AddSportsService {
-    @POST("/api/v1/users/me/profiles")
-    suspend fun addSportProfile(
-        @Body request: AddSportProfileRequest
-    ): BaseResponse<String?> // data가 null로 옴
-}
-
