+++ app/src/main/java/com/smashing/app/data/remote/service/MyService.kt
@@ -1,11 +1,13 @@
 package com.smashing.app.data.remote.service
 
 import com.smashing.app.data.remote.dto.BaseResponse
-import com.smashing.app.data.remote.dto.profile.my.MyPageData
-import com.smashing.app.data.remote.dto.profile.my.MyProfileReviewListData
-import com.smashing.app.data.remote.dto.profile.my.MyProfileSwitchRequest
+import com.smashing.app.data.remote.dto.my.AddSportProfileRequest
+import com.smashing.app.data.remote.dto.my.MyPageData
+import com.smashing.app.data.remote.dto.my.MyProfileReviewListData
+import com.smashing.app.data.remote.dto.my.MyProfileSwitchRequest
 import retrofit2.http.Body
 import retrofit2.http.GET
+import retrofit2.http.POST
 import retrofit2.http.PUT
 import retrofit2.http.Query
 
@@ -23,4 +25,9 @@ interface MyService {
     suspend fun putActiveMyProfile(
       @Body request: MyProfileSwitchRequest
     ): BaseResponse<Unit?>
+
+    @POST("/api/v1/users/me/profiles")
+    suspend fun addSportProfile(
+        @Body request: AddSportProfileRequest
+    ): BaseResponse<String?>
 }
