--- app/src/main/java/com/smashing/app/data/repository/api/AddSportsRepository.kt
@@ -1,7 +1,0 @@
-package com.smashing.app.data.repository.api
-
-import com.smashing.app.data.model.addsports.AddSportsInfo
-
-interface AddSportsRepository {
-        suspend fun addSportsProfile(info: AddSportsInfo): Result<Unit>
-}
