+++ app/src/main/java/com/smashing/app/data/repository/impl/MyRepositoryImpl.kt
@@ -3,11 +3,13 @@ package com.smashing.app.data.repository.impl
 import com.smashing.app.core.util.suspendRunCatching
 import com.smashing.app.data.mapper.toGameReviewPage
 import com.smashing.app.data.mapper.toMyPageInfo
+import com.smashing.app.data.mapper.toRequest
+import com.smashing.app.data.model.addsports.AddSportsInfo
 import com.smashing.app.data.model.cursor.CursorPage
 import com.smashing.app.data.model.profile.MyPageInfo
 import com.smashing.app.data.model.review.GameReview
 import com.smashing.app.data.remote.datasource.api.MyRemoteDataSource
-import com.smashing.app.data.remote.dto.profile.my.MyProfileSwitchRequest
+import com.smashing.app.data.remote.dto.my.MyProfileSwitchRequest
 import com.smashing.app.data.remote.dto.requireData
 import com.smashing.app.data.remote.service.MyService
 import com.smashing.app.data.repository.api.MyRepository
@@ -39,5 +41,10 @@ class MyRepositoryImpl @Inject constructor(
                 .putActiveMyProfile(MyProfileSwitchRequest(profileId = profileId))
                 .requireData()
         }
+    }
+
+    override suspend fun addSportsProfile(info: AddSportsInfo): Result<Unit> = suspendRunCatching {
+        myRemoteDataSource.addSportProfile(info.toRequest())
+
     }
 }
